refactor: dataInit 패키지 리팩토링 (#205)
