+++ src/main/java/ku_rum/backend/global/dataInit/CollegeInitializer.java
@@ -1,4 +1,4 @@
-package ku_rum.backend.global.dataInit2;
+package ku_rum.backend.global.dataInit;
 
 import ku_rum.backend.domain.college.domain.College;
 import org.springframework.stereotype.Component;
